chore: extract mods and functions
